Include error message for Forbidden responses

+++ test/functional/in/reeltime/reel/ReelFunctionalSpec.groovy
@@ -190,7 +190,7 @@ class ReelFunctionalSpec extends FunctionalSpec {
         def response = delete(request)
 
         then:
-        responseChecker.assertStatusCode(response, 403)
+        responseChecker.assertUnauthorizedError(response)
 
         cleanup:
         reelTimeClient.deleteReel(writeToken, reelId)
@@ -208,7 +208,7 @@ class ReelFunctionalSpec extends FunctionalSpec {
         def response = post(request)
 
         then:
-        responseChecker.assertStatusCode(response, 403)
+        responseChecker.assertUnauthorizedError(response)
     }
 
     void "list reels"() {
